Fix bad import in LoginForm tests; Rename app/services to app/api
