+++ src/features/auth/components/__tests__/RegisterForm.test.tsx
@@ -1,5 +1,5 @@
 import { userConstraints } from '@/lib/constants';
-import '@/mocks/firebaseMock';
+import '@/mocks/services/firebaseMock';
 import { userWithoutReviewMock as userMock } from '@/mocks';
 import renderWithProviders from '@/utils/renderWithProviders';
 import { screen } from '@testing-library/react';
